Fix CORS preflight requests failing: - Prevent proxy servlet to answer request before finding a match in the configuration
